# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Steward SaaS API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface Tier {
  id: string;
  name: string;
  description: string;
  price: number;
  features: string[];
  /** @nullable */
  stripePriceId: string | null;
  highlight?: boolean;
}

export interface CreateCheckoutRequest {
  tierId: string;
}

export interface CheckoutSessionResponse {
  url: string;
}

export interface PortalSessionResponse {
  url: string;
}

export interface SubscriptionStatusResponse {
  hasSubscription: boolean;
  /** @nullable */
  tierId?: string | null;
  /** @nullable */
  tierName?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  currentPeriodEnd?: string | null;
  /** @nullable */
  stripeCustomerId?: string | null;
}

export interface Organization {
  id: string;
  userId: string;
  name: string;
  type: string;
  /** @nullable */
  category: string | null;
  /** @nullable */
  tier: string | null;
  /** @nullable */
  stripeCustomerId?: string | null;
  /** @nullable */
  stripeSubscriptionId?: string | null;
  /** @nullable */
  subscriptionStatus?: string | null;
  /** @nullable */
  slug?: string | null;
  createdAt: string;
}

export interface OrganizationResponse {
  organization: Organization | null;
}

export interface UpsertOrganizationRequest {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  type: string;
  category?: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  /**
   * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
   */
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type ListTiers200 = {
  tiers: Tier[];
};