# lib/api-zod/src/generated/types/logoutSuccess.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Steward SaaS API specification
 * OpenAPI spec version: 0.1.0
 */

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;