# lib/api-zod/src/generated/types/organization.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Steward SaaS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Organization {
  id: string;
  userId: string;
  name: string;
  type: string;
  /** @nullable */
  category: string | null;
  /** @nullable */
  tier: string | null;
  /** @nullable */
  stripeCustomerId?: string | null;
  /** @nullable */
  stripeSubscriptionId?: string | null;
  /** @nullable */
  subscriptionStatus?: string | null;
  /** @nullable */
  slug?: string | null;
  createdAt: string;
}