# lib/api-zod/src/generated/types/organizationResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Steward SaaS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Organization } from "./organization";

export interface OrganizationResponse {
  organization: Organization | null;
}