# lib/api-zod/src/generated/types/subscriptionStatusResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Steward SaaS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubscriptionStatusResponse {
  hasSubscription: boolean;
  /** @nullable */
  tierId?: string | null;
  /** @nullable */
  tierName?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  currentPeriodEnd?: string | null;
  /** @nullable */
  stripeCustomerId?: string | null;
}