# lib/api-zod/src/generated/types/upsertOrganizationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Steward SaaS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpsertOrganizationRequest {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  type: string;
  category?: string;
}